pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = "Abstract-assessment"
authors = ["Iván Abregú"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
psycopg2 = "^2.8.5"
django-cors-headers = "^3.4.0"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-filter = "^2.3.0"
requests = "^2.28.2"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
